refactor: 공통 hook 분리, schema/option 분리, 코드 스타일 변경
