+++ src/hooks/useGetLoggedInUserId.ts
@@ -1,0 +1,9 @@
+import { useGetLoggedInUserApi } from '@/apis/loggedInUser';
+
+const useGetLoggedInUserId = () => {
+  const { id: currentUserId } = useGetLoggedInUserApi();
+
+  return currentUserId;
+};
+
+export default useGetLoggedInUserId;
