+++ src/pages/GatheringUnfix/hooks/useGatheringUnfix.ts
@@ -1,7 +1,4 @@
-import {
-  type GatheringUnfixRequest,
-  useDeleteGatheringUnfixApi,
-} from '@/apis/deleteGatheringUnfix';
+import { useDeleteGatheringUnfixApi } from '@/apis/gatheringUnfix';
 import { showErrorToast } from '@/utils/showToast';
 
 import type { OnGatheringUnfix } from '../components/GatheringUnfixForm';
@@ -9,8 +6,11 @@ import type { OnGatheringUnfix } from '../components/GatheringUnfixForm';
 const useGatheringUnfix = (onUnfix: OnGatheringUnfix, gatheringId: string) => {
   const gatheringUnfixApi = useDeleteGatheringUnfixApi(gatheringId);
 
-  return async (request: GatheringUnfixRequest) => {
-    const { data, isBadRequest } = await gatheringUnfixApi(request);
+  return async (reason: string) => {
+    const { data, isBadRequest } = await gatheringUnfixApi({
+      reason,
+      gatheringId,
+    });
 
     if (isBadRequest) {
       return showErrorToast(data.message);
