+++ src/pages/GatheringUnfix/index.tsx
@@ -2,35 +2,33 @@ import { useEffect, useState } from 'react';
 
 import { useNavigate, useParams } from 'react-router-dom';
 
-import { useGetGatheringDetailApi } from '@/apis/getGatheringDetail';
-import { useGetLoggedInUserApi } from '@/apis/getLoggedInUser';
+import { useGetGatheringDetailApi } from '@/apis/gatheringDetail';
 import Modal from '@/components/Modal';
 import TabBar from '@/components/TabBar';
 import Alert from '@/components/alert';
 import { UNFIX_GATHERING_ALERT_MESSAGE } from '@/constants/messages/alert';
 import { SUCCESS_UNFIX_GATHERING_MODAL_MESSAGE } from '@/constants/messages/modal';
+import { GATHERINGS_PAGE_URL } from '@/constants/pageRoutes';
+import useGetLoggedInUserId from '@/hooks/useGetLoggedInUserId';
 
 import GatheringUnfixForm from './components/GatheringUnfixForm';
 
 const GatheringUnfixPage = () => {
   const { gatheringId } = useParams();
 
-  if (!gatheringId) {
-    throw new Error('gathering id is required');
-  }
-
   const navigate = useNavigate();
   const [isModalOpen, setIsModalOpen] = useState(false);
 
-  const { id: currentUserId } = useGetLoggedInUserApi();
-  const { isFix, participantResponse: participants } =
-    useGetGatheringDetailApi(gatheringId);
+  const currentUserId = useGetLoggedInUserId();
+  const { isFix, participantResponse: participants } = useGetGatheringDetailApi(
+    gatheringId as string,
+  );
 
   const handleOpenModal = () => setIsModalOpen(true);
 
   const handleCloseModal = () => {
     setIsModalOpen(false);
-    navigate(`/gatherings/${gatheringId}`);
+    navigate(`${GATHERINGS_PAGE_URL}/${gatheringId}`);
   };
 
   useEffect(() => {
@@ -42,7 +40,7 @@ const GatheringUnfixPage = () => {
       throw new Error('Inaccessible page');
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [currentUserId, participants]);
+  }, []);
 
   return (
     <div className='flex h-full flex-col'>
